Retitle line moving as viewport adjusting
Like Nano, we want to be able to adjust the viewport independently
of the cursor.  I've experimented with keeping the cursor centered
in the viewport at all times (which makes it easy to find!)
but sometimes that puts the wrong amount of context onscreen.
It's more user-friendly to let that scroll seperately.