Preliminary cleanup of buffer switching
This is still broken because we can't reuse the window offset
when switching to a new buffer because that could be outside the
buffer's range, but resetting to 0 each time isn't right either.

The hard part here is working out how it's *should* behave.
I probably want to keep the old offsets on a pane-by-pane basis.