Rename project
My initial conception was that of a "Rust Nano".
But on further reflection, my goal is to be something
a little more featureful than Nano, so Micro would be more
appropriate.

But of course that name is taken on Crates.io, so
I'll use micro-editor as the actual crate name.