Eliminate manual viewport control
Being able to alt-up/down to adjust the viewport
and Ctrl-L to center it on the cursor are Nano-isms
that consumes a lot of keybindings.

Trouble is, when bouncing through a document - either
by jumping to a line or doing a search - I wind up playing
a game of hunt-the-cursor which is irritating.

By centering the cursor vertically at all times whenever
possible, it's trivial to locate because my eyes always
stay in the center of my screen.