Fix how the modified flag is calculated
How it should behave in practice is a bit different from how
it's perceived.  If we modify a file, then make more modifications
that put it back to its original state, the file should no longer
be flagged as modified at that point.

The only way to really make sure this works in all circumstances
is to put rope modifications behind a handle which checks
whether there's been any modifications at drop time.
We just lend a handle which DerefMuts to a rope,
let the caller do whatever, then run a comparison.